+++ extensions/FeatureBasedAlignment.py
@@ -14,6 +14,11 @@ class FeatureBasedAlignment(ImagePreprocessor):
         self.ref_img = cv2.imread(self.ref_path, cv2.IMREAD_GRAYSCALE)
         self.MAX_FEATURES = options.get('MaxFeatures', MAX_FEATURES)
         self.GOOD_MATCH_PERCENT = options.get('GoodMatchPercent', GOOD_MATCH_PERCENT)
+        self.TRANSFORM_2D = options.get('2D', False)
+        # Extract keypoints and description of source image
+        self.orb = cv2.ORB_create(self.MAX_FEATURES)
+        self.to_keypoints, self.to_descriptors = self.orb.detectAndCompute(self.ref_img, None)
+
 
     def __str__(self):
         return self.ref_path
@@ -32,13 +37,11 @@ class FeatureBasedAlignment(ImagePreprocessor):
         img = cv2.normalize(img, 0, 255, norm_type=cv2.NORM_MINMAX)
         
         # Detect ORB features and compute descriptors.
-        orb = cv2.ORB_create(self.MAX_FEATURES)
-        keypoints1, descriptors1 = orb.detectAndCompute(img, None)
-        keypoints2, descriptors2 = orb.detectAndCompute(self.ref_img, None)
+        from_keypoints, from_descriptors = self.orb.detectAndCompute(img, None)
         
         # Match features.
         matcher = cv2.DescriptorMatcher_create(cv2.DESCRIPTOR_MATCHER_BRUTEFORCE_HAMMING)
-        matches = matcher.match(descriptors1, descriptors2, None)
+        matches = matcher.match(from_descriptors, self.to_descriptors, None)
         
         # Sort matches by score
         matches.sort(key=lambda x: x.distance, reverse=False)
@@ -48,7 +51,7 @@ class FeatureBasedAlignment(ImagePreprocessor):
         matches = matches[:numGoodMatches]
         
         # Draw top matches
-        #imMatches = cv2.drawMatches(im1, keypoints1, im2, keypoints2, matches, None)
+        #imMatches = cv2.drawMatches(im1, from_keypoints, im2, self.to_keypoints, matches, None)
         #show('Aligning', imMatches, resize=True)    
         
         # Extract location of good matches
@@ -56,14 +59,16 @@ class FeatureBasedAlignment(ImagePreprocessor):
         points2 = np.zeros((len(matches), 2), dtype=np.float32)
         
         for i, match in enumerate(matches):
-            points1[i, :] = keypoints1[match.queryIdx].pt
-            points2[i, :] = keypoints2[match.trainIdx].pt
+            points1[i, :] = from_keypoints[match.queryIdx].pt
+            points2[i, :] = self.to_keypoints[match.trainIdx].pt
         
         # Find homography
-        h, mask = cv2.findHomography(points1, points2, cv2.RANSAC)
-        
-        # Use homography
         height, width = self.ref_img.shape
-        im1Reg = cv2.warpPerspective(img, h, (width, height))
+        if self.TRANSFORM_2D:
+            m, inliers = cv2.estimateAffine2D(points1, points2)
+            return cv2.warpAffine(img, m, (width, height))
+        else:        
+            # Use homography
+            h, mask = cv2.findHomography(points1, points2, cv2.RANSAC)
+            return cv2.warpPerspective(img, h, (width, height))
         
-        return im1Reg
